requirements.txt
PyAudio
SpeechRecognition
openai-whisper
wheel
playsound